Rename _samples suffix to _values
